override getColumnCanGlobalFilter if columnDef.enableGlobalFilter explicitly set to true
table.options.getColumnCanGlobalFilter will return "false" if the value type is not a string or a number. in instances where the value type is not a primitive, (i.e., an array or object) and there is a custom global filter function defined to handle the value type, the column is still not passed to the global filter function because this createColumn logic sequence still resolves to false. This allows an override of the getColumnCanGlobalFilter if the user explicitly states in the column def that the enableGlobalFilter is set to true